requirements.dev.txt
-r requirements.txt
django-debug-toolbar==1.2.1
uwsgi
django-extensions==1.3.8
watchdog